Refactor slippage min amount workflow: removed unnecessary nodes, updated input schema, and improved output handling
